backend/requirements.txt
# ─── Core API ───────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1

# ─── Deep Learning ──────────────────────────
tensorflow==2.16.1       # or tensorflow-cpu==2.16.1 for CPU-only
# torch==2.3.0            # uncomment if using PyTorch version instead
# torchvision==0.18.0

# ─── Image Processing ───────────────────────
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# ─── Data / Evaluation ──────────────────────
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2

# ─── Utilities ──────────────────────────────
tqdm==4.66.4
python-dotenv==1.0.1
